requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bcolz-zipline==1.2.9
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filterpy==1.4.5
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
graphviz==0.20.3
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.3.post0
matplotlib==3.9.1
menpo==0.11.1
multidict==6.0.5
mxnet==1.2.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prettytable==3.10.2
protobuf==5.27.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tifffile==2024.7.2
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==1.4.0.post0
torchvision==0.14.1+cu117
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
zipp==3.19.2